Move models from __init__.py to models module
`__init__.py` shouldn't have that many models and make the file large. I
moved (most of) the models from there to the `models` module.
