+++ compass_sdk/clients/rbac.py
@@ -5,7 +5,7 @@ import requests
 from pydantic import BaseModel
 from requests import HTTPError
 
-from compass_sdk.models.rbac import (
+from compass_sdk.models import (
     GroupCreateRequest,
     GroupCreateResponse,
     GroupDeleteResponse,
